EWPP-441: TextFeaturedMediaAssert.
